requirements.txt
Django==3.0.4
biopython
pickle
collection